Initial Commit: Basic Operations Ribbon
This commit adds a basic, non-functional operations ribbon to the Show
Design Tool.  Each subset of operations is packaged off in its own
widget to allow for easy modification, addition, or removal of similar
types of operations.  The operations ribbon also allows for resizing of
the window by updating the ribbon to 1 of 3 possible configurations, and
even dynamically adjusts for added or removed buttons or subsets of
buttons.