Add Show Notebook
This commit adds a show notebook to the Show Design Tool layout.
This show notebook contains the individual shows for the current
global show that is being edited.  Additionally, the root window
for the Show Design Tool allows for the notebook to expand to take
ANY extra vertical space as the window resizes.

Limits are put in place when calculating the minimum size for the
overall tool to make sure that it does not extend of the screen when
the global show is initialized with a large number of sets or
individual shows.

Lastly, each individual show effectively scrolls in tandem; in other
words, when switching between tabs of individual shows, the sets that
are being viewed will stay consistent window to window, to save the user
time when editing the same set in multiple shows.