Fix Operations Ribbon Layout Bug
This commit remedies the issue that occurs when the show notebook ends
up being large enough to dominate the horizontal space (in addition to
the vertical space) of the overall tool.  When this occured, the
operations ribbon would expand to the larger window size, causing its
own layout size requirements to be skewed.

This problem is fixed by setting the operations ribbon to expand AFTER
it has determined its layout size requirements, allowing the
measurements to be taken on the absolute minimum size of the ribbon.