Resolve Lag with High Show Count
This commit addresses the lag that is generated when a large show count
is used.  When using the notebook, the tabs not only become impossible
to read if the show count skyrockets, but the entire Show Design Tool is
barely responsive.

This new solution makes use of a Spinbox and Label instead, by
populating only ONE Treeview with the values of the currently selected
individual show (shown in the Spinbox).  This reduces the
SDTIndividualShow class to a wrapper for show instructions, and
replaces the SDTShowNotebook class with the SDTShowFrame class that will
contain the Spinbox, Label, and Treeview.

While SDTShowFrame now contains a Spinbox, Label, and Treeview, the
Spinbox and Treeview are now SDTShowSpinbox and SDTShowTreeview classes
of their own to separate out and convey clearer, easier to parse
functionality and code.