Prepare for Button Functionality Implementation
This commit changes the current scheme of passing updates to the Show
Design Tool window to output on the status label and store in the status
log.  Instead of relying on the current structure/nesting of the
windows, each class prefixed with SDT will receive the ShowDesignTool
class as an input (similar to how master is passed).  This allows any
messages or notifications to be passed to the tool as a sort of
"hub", to request information from a central point, as opposed to
relying on the structure to the root frame.

Additionally, redundant self.master assignments were removed, and all
uses of self.sdt are now checked to ensure they are not NoneType objects
before using them as intended.

Variables were also renamed to be more descriptive and less
confusing, and show count and set count variables were added to the main
Show Design Tool class to prepare for using such variables in file
opening and saving actions.

Lastly, this commit creates an empty bounds file to be filled with
"macros" that can be used in the various operation classes to limit
the behavior of the operations appropriately.