File Operations: Save and Open
This commit implements the necessary functions for storing the global
show (and all child data) in a human readable json file (either directly
in a .json container or into a .zip archive). These simply "bubble
down" through the Show Design Tool hierarchy and bubble back up with the
necessary dictionaries. This overall dictionary is then converted to a
json string and stored off into the appropriate file format.

The necessary functions are also implemented to open and parse a file,
and then propogate it throughout the Show Design Tool structure.
The SDTFileOps class handles the verification of the file to ensure that
the structure is correctly formatted as the design tool structure
expects it. It also verifies that the show counts and set counts match
all of the show and set instruction lists so that no out of bounds
exceptions are possible. Given the sufficient rigor of this
verification, all of the "reconfigure" instructions in the design tool
structure assume that the dictionary is formatted properly and simply
plugs the information directly in, simplifying the readability of that
code.