Show Operations: Add and Remove Shows
First, this commit adds the bounds necessary for adding and
removing individual shows (the operations that are supported).

Next, this commits implements the necessary functions to accomplish
adding a new individual show to the current global show. Upon the new
addition of the show, it is populated with the appropriate "Lights Off"
set instructions (RGB of 0, 0, 0).  And the show is appended as
the last individual show in the global show.  But, if the maximum
show count is reached, an error dialog pops up and the show is not
added.

Lastly, this commits implements the necessary functions to accomplish
removing an individual show from the current global show.  Currently, the
operation removes the show that is currently selected via the show
spinbox (as this seems most intuitive), and all of the shows after the
removed show are decremented  to fill the "hole" left by the removed
show. If the minimum show count is reached, a warning dialog pops up
to see if the user wishes to reset the only remaining show to
"Lights Off".