Set Operations: Add and Remove Functionality
This commits implements the necessary functions to accomplish adding and
removing sets to the current global show.

Upon the new addition of a set, it is populated with "Lights Off" set
instructions (all 0s) for all individual shows.  Additionally, the added
set is appended as to the end of every individual show.  If the maximum
number of sets is reached, an error message pops up.

When removing a set, the currently selected set will be removed, and
all sets after the removed set will be decremented appropriately.  If a
set is not selected, the last set for each individual show will be
removed. If the minimum number of sets is reached, an error message
pops up.