Set Instruction Ops: Edit and Clear Functionality
This commits implements the necessary functions to accomplish editing
and clearing set instructions, which with the current functionality
simply means picking colors and setting to "Lights Off".

If a set instruction is selected, that set instruction will be edited or
cleared. If a set instruction is not selected, an error message pops up
prompting the user to select a set instruction.