Pack Operations: Laying Groundwork
This commit adds a skeleton for the Pack Operations that will be added
in future commits.  First, it establishes the hierarchy of calls from
the operation ribbon itself to the necessary classes of the Show Design
Tool.  It also integrates the portion of the window that will be
dedicated to the Pack IDs and all relevant information, along with the
Program Button.  Currently, the Pack information only consists of the
Pack ID, and any added information will be integrated as necessary
in future commits.