Pack Operations: Pack ID Add, Remove, and File
This commit adds a prompt for adding a pack ID, as well as the
actual implementation necessary to store off Pack ID information.
The prompt will validate any inputs, and the ID is verified before
being stored off into the list.  Additionally, no duplicate Pack IDs
are permitted to be added until the other is removed from the
list.

Pack IDs are also able to be removed from the current Global Show
list (deleted the selected one if any are highlighted, deleted the
last if none are highlighted).

Lastly, this information has now been added to the global show
file json format (under the key "pack_info_list").  The necessary
verification has been added appropriately to support opening a
file with Pack IDs, and the UI will populate the list accordingly.
The proper information is also stored off during a save action.