Pack Operations: Pack ID Aliasing
This commit adds a prompt for aliasing a pack ID, as well as the
actual implementation necessary to store off alias along with the pack
information. The pack alias consists of a single ASCII printable and
non-whitespace character, and a number between 1 and 99 (inclusive).
If the user attempts to alias multiple pack IDs with the same alias,
an error message will pop up and the duplicate alias will not be
permitted.