Pack Operations: Pack ID Assignments
This commit adds a prompt for assigning pack IDs to individual shows,
as well as the actual implementation necessary to store off the show
number along with the pack information. The Pack ID will be assigned
to one of the show numbers presently available, and if that show
is deleted, all packs assigned to that show will be unassigned and
the other show assignment numbers will be adjusted accordingly.